bump to new equistore
