+++ setup.cfg
@@ -17,7 +17,6 @@ classifiers =
     Operating System :: Microsoft :: Windows
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -32,7 +31,7 @@ keywords =
     Science
     Machine Learning
 
-python_requires = >=3.6
+python_requires = >=3.7
 
 project_urls =
     Source = https://github.com/lab-cosmo/equisolve
@@ -46,7 +45,7 @@ packages = find:
 include_package_data = True
 
 install_requires =
-    equistore @ https://github.com/lab-cosmo/equistore/archive/ee3077b.zip
+    equistore @ https://github.com/lab-cosmo/equistore/archive/080e44a.zip
     numpy
     scipy
