+++ tests/equisolve_tests/init.py
