+++ tests/equisolve_tests/numpy/models/linear_model.py
