+++ tests/equisolve_tests/utils/convert.py
