+++ tests/equisolve_tests/utils/metrics.py
