+++ tests/equisolve_tests/utils/split_data.py
