+++ tox.ini
@@ -12,35 +12,15 @@ lint_folders = {toxinidir}/examples {toxinidir}/src {toxinidir}/tests {toxinidir
 
 [testenv:tests]
 # this environement runs Python tests
-
-setenv =
-    PYTHONPATH={toxinidir}/tests
-    PYTHONUNBUFFERED=yes
 usedevelop = true
-# installs dependencies we need for testing
-# by using tox the developer don't need to manage this dependencies
-# him/herself
 deps =
     ase
-    coverage[toml]
-    hypothesis
     pytest
-    pytest-cov
-# before running the tests erases any prerecord of coverage
-commands_pre =
-    coverage erase
-# execute pytest
 commands =
-    pytest --cov --cov-report=term-missing --cov-append --cov-config=.coveragerc -vv --hypothesis-show-statistics {posargs}
-# after executing the pytest assembles the coverage reports
-commands_post =
-    coverage report
-    coverage html
-    coverage xml
+    pytest
+
 
 [testenv:lint]
-# this environement lints the Python code with flake8 (code linter), black (code
-# formatter), and isort (sorting of imports)
 skip_install = true
 deps =
     flake8
